refactor: use Application base class
